# 05_API/openapi_core.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS desktop core command API",
    "version": "1.0.0",
    "description": "Hợp đồng thiết kế lõi, chưa là API server đã triển khai. CRUD danh mục theo conventions trong README."
  },
  "servers": [
    {
      "url": "https://wms.example.internal/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/purchase-orders": {
      "get": {
        "operationId": "listReceivablePurchaseOrders",
        "summary": "UC06 list receivable purchase orders",
        "description": "Only APPROVED or PARTIAL purchase orders with positive remaining quantity that the caller may read under document.read: assigned warehouse plus own/assigned document for RECEIVER. Return only the minimal fields in ReceivablePurchaseOrder. Remaining is derived from net posted receipts, not drafts. Cursor is tied to a server-maintained as_of snapshot and immutable (created_at,id) order so intervening edits cannot duplicate or skip rows.",
        "x-permission": "document.read",
        "x-use-case": "UC06",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/AsOf"
          }
        ],
        "responses": {
          "200": {
            "description": "Receivable purchase orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderPage"
                },
                "example": {
                  "items": [
                    {
                      "id": "c0000000-0000-4000-8000-000000000001",
                      "number": "PO-001",
                      "warehouse_id": "10000000-0000-4000-8000-000000000001",
                      "created_at": "2026-10-01T09:00:00+07:00",
                      "status": "APPROVED",
                      "lines": [
                        {
                          "source_line_id": "20000000-0000-4000-8000-000000000001",
                          "product_id": "30000000-0000-4000-8000-000000000001",
                          "uom_id": "40000000-0000-4000-8000-000000000001",
                          "ordered_base": "100.000000",
                          "remaining_base": "100.000000"
                        }
                      ]
                    }
                  ],
                  "as_of": "2026-10-02T09:00:00+07:00",
                  "next_cursor": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          }
        }
      }
    },
    "/purchase-orders/{id}": {
      "get": {
        "operationId": "getReceivablePurchaseOrder",
        "summary": "UC06 get purchase order remaining quantities",
        "description": "Require document.read on this PO, including own/assigned document for RECEIVER and assigned warehouse. Return current net posted remaining per source line; recheck under lock when posting.",
        "x-permission": "document.read",
        "x-use-case": "UC06",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivablePurchaseOrder"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/openings/{id}/post": {
      "post": {
        "operationId": "postOpening",
        "summary": "UC23 postOpening",
        "description": "Quyền: opening.post. Kiểm tra scope/state ở server. Quantity không cho phép thay đổi nguồn/đích tùy ý ngoài rule loại phiếu.",
        "x-permission": "opening.post",
        "x-use-case": "UC23",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã commit hoặc replay kết quả",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tồn tại hoặc không được phép thấy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version/idempotency/state/stock conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Tạm thời không phục vụ, tra operation trước retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/approval-requests/{id}/decide": {
      "post": {
        "operationId": "decideApproval",
        "summary": "UC05,UC18 decideApproval",
        "x-receiving-example": {
          "request": {
            "expected_version": 2,
            "decision": "APPROVE"
          },
          "response": {
            "id": "f0000000-0000-4000-8000-000000000001",
            "status": "APPROVED",
            "version": 3,
            "request_id": "b0000000-0000-4000-8000-000000000008"
          }
        },
        "description": "Quyền: document.approve OR adjustment.approve OR opening.approve by kind. Kiểm tra scope/state ở server. Quantity không cho phép thay đổi nguồn/đích tùy ý ngoài rule loại phiếu.",
        "x-permission": "document.approve OR adjustment.approve OR opening.approve by kind",
        "x-use-case": "UC05,UC18",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đã commit hoặc replay kết quả",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tồn tại hoặc không được phép thấy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version/idempotency/state/stock conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Tạm thời không phục vụ, tra operation trước retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/operations/{key}": {
      "get": {
        "operationId": "getOperation",
        "description": "Read the immutable committed-operation projection by the original Idempotency-Key. operation_status=COMMITTED describes the command, while status is the resource state. id/status/version are copied from the committed command result; request_id is the original command's stored request ID, not this GET's trace ID. transaction_id is present for posting. A 200 proves commit; 404 is inconclusive. To recover the exact original HTTP status/body, replay the same method, route, body and Idempotency-Key (also the same execution_key for posting). Recheck current resource permission before returning data. Current GET /receipts/{id} may show later state and is not an exact replay.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable summary of a committed command; original response is obtained by same-key replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationLookupResult"
                },
                "examples": {
                  "createReceipt": {
                    "summary": "Draft creation after a lost 201 response",
                    "value": {
                      "operation_status": "COMMITTED",
                      "id": "90000000-0000-4000-8000-000000000001",
                      "status": "DRAFT",
                      "version": 1,
                      "request_id": "b0000000-0000-4000-8000-000000000003"
                    }
                  },
                  "patchReceipt": {
                    "summary": "Draft edit after a lost 200 response",
                    "value": {
                      "operation_status": "COMMITTED",
                      "id": "90000000-0000-4000-8000-000000000001",
                      "status": "DRAFT",
                      "version": 2,
                      "request_id": "b0000000-0000-4000-8000-000000000004"
                    }
                  },
                  "reviseReceipt": {
                    "summary": "Approved receipt returned to DRAFT",
                    "value": {
                      "operation_status": "COMMITTED",
                      "id": "90000000-0000-4000-8000-000000000001",
                      "status": "DRAFT",
                      "version": 4,
                      "request_id": "b0000000-0000-4000-8000-000000000005"
                    }
                  },
                  "postReceipt": {
                    "summary": "Receipt posting after a lost 200 response",
                    "value": {
                      "operation_status": "COMMITTED",
                      "id": "90000000-0000-4000-8000-000000000001",
                      "status": "COMPLETED",
                      "version": 4,
                      "transaction_id": "a0000000-0000-4000-8000-000000000001",
                      "request_id": "b0000000-0000-4000-8000-000000000001"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "No visible committed result; request may still have executed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "OPERATION_NOT_FOUND",
                  "message": "No committed result is currently visible for this key.",
                  "request_id": "b0000000-0000-4000-8000-000000000009",
                  "retryable": true
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 100
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Opaque next_cursor from the preceding response; do not construct or alter it."
      },
      "AsOf": {
        "name": "as_of",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "Optional on first page; use returned as_of for subsequent pages. UTC offset required."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Visible resource but action denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Missing resource or outside caller scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InvalidInput": {
        "description": "Invalid field, UOM, tracking or quantity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "INVALID_QUANTITY",
              "message": "Quantity must be positive and exact in base UOM.",
              "request_id": "b0000000-0000-4000-8000-000000000001",
              "retryable": false,
              "field_errors": [
                {
                  "line_id": "60000000-0000-4000-8000-000000000001",
                  "field": "quantity",
                  "code": "INVALID_QUANTITY",
                  "message": "Use a positive decimal string with at most 6 places."
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
      "PositiveQuantity": {
        "type": "string",
        "pattern": "^(?=.*[1-9])\\d{1,14}(\\.\\d{1,6})?$",
        "description": "Positive numeric(20,6) decimal string; never JSON number. Server checks UOM precision and exact base conversion."
      },
      "NonnegativeQuantity": {
        "type": "string",
        "pattern": "^\\d{1,14}(\\.\\d{1,6})?$",
        "description": "Nonnegative numeric(20,6) base-UOM decimal string."
      },
...
      "PurchaseOrderLineRemaining": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source_line_id",
          "product_id",
          "uom_id",
          "ordered_base",
          "remaining_base"
        ],
        "properties": {
          "source_line_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "uom_id": {
            "type": "string",
            "format": "uuid"
          },
          "ordered_base": {
            "$ref": "#/components/schemas/PositiveQuantity"
          },
          "remaining_base": {
            "$ref": "#/components/schemas/NonnegativeQuantity"
          }
        }
      },
      "ReceivablePurchaseOrder": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "number",
          "warehouse_id",
          "created_at",
          "status",
          "lines"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "string"
          },
          "warehouse_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "PARTIAL"
            ]
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderLineRemaining"
            }
          }
        }
      },
      "PurchaseOrderPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "as_of",
          "next_cursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivablePurchaseOrder"
            }
          },
          "as_of": {
            "type": "string",
            "format": "date-time"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Command": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version",
          "execution_key",
          "lines"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "execution_key": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "maxItems": 200,
            "items": {
              "$ref": "#/components/schemas/ExecutionLine"
            }
          }
        }
      },
      "ExecutionLine": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "document_line_id",
          "stock_item_id",
          "quantity_base"
        ],
        "properties": {
          "document_line_id": {
            "type": "string",
            "format": "uuid"
          },
          "stock_item_id": {
            "type": "string",
            "format": "uuid"
          },
          "source_location_id": {
            "type": "string",
            "format": "uuid"
          },
          "destination_location_id": {
            "type": "string",
            "format": "uuid"
          },
          "reservation_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity_base": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,6})?$",
            "description": "Decimal chuỗi >0, server kiểm tra độ chính xác UOM; SERIAL=1."
          }
        }
      },
      "Decision": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expected_version",
          "decision"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "REJECT"
            ]
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "retryable": {
            "type": "boolean"
          },
          "field_errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "line_id": {
                  "type": "string"
                },
                "field": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Result": {
        "type": "object",
        "required": [
          "id",
          "status",
          "request_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "description": "Có với document/count, bỏ trống với import job."
          },
          "status": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "OperationLookupResult": {
        "description": "Immutable projection of a committed command; status is resource state, not operation state.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "required": [
              "operation_status"
            ],
            "properties": {
              "operation_status": {
                "type": "string",
                "enum": [
                  "COMMITTED"
                ]
              }
            }
          }
        ]
      },
...
    }
  }
}